requirements.txt
sentence-transformers==5.0.0
datasets
transformers
torch
accelerate
matplotlib
python-dotenv